# backend/openapi-schema.yml
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
...
  /api/reports/fees/:
    get:
      operationId: reports_fees_retrieve
      tags:
      - reports
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/students/:
    get:
      operationId: students_list
      tags:
      - students
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
          description: ''
    post:
      operationId: students_create
      tags:
      - students
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Student'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Student'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: ''
...
  /api/students/{id}/fees/:
    get:
      operationId: students_fees_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Student.
        required: true
      tags:
      - students
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: ''
  /api/students/import/:
    post:
      operationId: students_import_create
      tags:
      - students
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Student'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Student'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: ''
  /api/teachers/:
    get:
      operationId: teachers_list
      tags:
      - teachers
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
          description: ''
    post:
      operationId: teachers_create
      tags:
      - teachers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Teacher'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Teacher'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
          description: ''
...
components:
  schemas:
...
    StatusEnum:
      enum:
      - active
      - inactive
      type: string
      description: |-
        * `active` - Active
        * `inactive` - Inactive
    Student:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        branch:
          type: integer
        branch_name:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 140
        parent_name:
          type: string
          maxLength: 140
        contact_number:
          type: string
          maxLength: 24
        address:
          type: string
        standard:
          type: string
          maxLength: 64
        batch_time:
          type: string
          maxLength: 64
        roll_number:
          type: string
          maxLength: 32
        admission_date:
          type: string
          format: date
          nullable: true
        decided_fee:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        paid_fee:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        fee_left:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        critical_notes:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - branch
      - branch_name
      - created_at
      - fee_left
      - id
      - name
      - updated_at
...
    Teacher:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        branch:
          type: integer
        name:
          type: string
          maxLength: 140
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          maxLength: 24
        subject:
          type: string
          maxLength: 120
        assigned_standard:
          type: string
          maxLength: 64
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - branch
      - created_at
      - id
      - name
      - updated_at
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT